fix : in-place modification bug in rotate_matrix kernel
rotate_matrix changed the state of passed in omega_collection making it
normalized (unit vector). This gets transmitted back to the user scope.
If the user reuses omega_collection in her/his code, then we would be
dealing with unit vectors in the direction of omega and not omega itself.
Thanks pytest!