Bug fix: mass computation
`mass` has to be computed on nodes instead of on elements. This commit
fixes `mass` computation and dimension `mass` array. `mass` computed by
rod constructor in `_rod.py` and it has a dimension of `n_element+1`

This commit also fixes test cases for `mass` array in
`test_rod_initialisation.py` .