bug fix: potential bug in `mass` computation
There was a possibility previous mass computation can cause problem for
elements, that has different volume, i.e. for tapered case. Now for each
element we first set `mass` for the first node of that element. Then we
set the `mass` for the second node of that element. Also, mass can be
interpreted using `density` and `volume`. Thus now it is computed in
`_CosseratRodBase` constructor, not in `classmethod`.