cleanup : make `collection` explicit in physical quantities
The name provided for `position`, `directors`, `velocity`, `omega`
was not descrpitive enough. I can think of two problems:

1. Why is `position` singular, while `directors` are plural? It makes
more sense if all are plural (as we store multiple element velocities
in the same np.ndarray). If some user needs and requests a velocity,
does she/he request a `velocity` or `velocities` : with the present
naming convention mixing singular and plurals, it is not obvious.

2. We should make explicit that we store MULTIPLE of these quantities
in one data-structure. By adding the `collection` suffix, I think
this becomes obvious. e.g. `position_collection` compared to `position`
and `director_collection` compared to `director`.
An added benefit
is that if we store a singular element, we can name it `position` or
`position_collection` depending upon our discretion---if in some part
of code we want to expose this information (how we store it as a
number and not arrays), we call it `position`, else `position_collection`.