bug fix: some indexing issues and improvements
This commit fixes bugs in `InteractionPlane` class, especially inside
`apply_normal_force` function. There were issues regarding the indexing
when computing `total_force` on elements. In order to solve that a new
function called `nodes_to_elements` is added. This function basically
takes forces acting on nodes and converts them to elemental forces.
Second bug was in `no_contact_pts` calculation. Here, effect of rod
thickness or radius was not considered, now with this commit we are
taking care of the rod thickness.
Finally, all dot products and outer products are converted to `einsum`
for more efficient code.