fix : rotation formulae that conforms to  row-wise director representation
A (major?) bug from the paper which propogates the minus sign
is because our rotation formulae is incorrect, and needs to have
an additional sign coming in as part of the skew-symmetry of the
cross-product operator. This is fixed in both the rotate and
log kernels in this commit.

Ask Teja for supplementary documentation if you find this hard
to understand