feat : small tolerances in rotate kernels
More specifically we add tolerances:
1. When normalizing axis_collection in _get_rotation_matrix. This
runs into trouble when the axis is 0.0 in all three-dimensions identically
(common use-case when starting with zero angular velocity)

So we normalize axis as axis/(norm + eps), to ensure there's no 0.0/0.0
divides

2. Trace calculation of the rotation matrix to get \theta according to
Tr = 1 + 2 \cos(\theta). While analytically the trace of a rotation
matrix \in [-1, 3], numerically this might be [-1-eps, 3+eps] in a worst
case scenario. This causes the `np.arccos` routine to return nans all over
the place.

To ensure trace belongs to range, we now do Tr - eps = 1 + 2 \cos(\theta).
This seems problematic when Tr=-1 at the first go, but rarely in our rods
does Tr = -1 (which implies adjacent elements are anti-parallel to one
another, which is a physically unrealistic scenario).

3. In x/\sin(x) calculation in inv_rotate. Once again when x is almost 0.0
we run into 0.0/0.0 errors. To avoid it, we put x/sin(x+eps). again this
is not problematic as x can NEVER be negative, which is a promise of the np.
arccos routine in step 2