fix : workflow in symplectic steppers
Earlier, all steps in the symplectic stepper were collected
together and looped over. However, this causes a problem with rods
which expects the boundary conditions, forcings etc. to only
appear at the end of the kinematic step (and not the dynamic step).

To distinguish these two-steps, we now split into two method lists:
one kinematic and one dynamic so that we apply forcings etc correctly.