bug fix: applying BC after Verlet dynamic step
This commit fixes an issue in timestepper. In previous case
`constraints` are called only after `kinematic` step is done and it was
not called after completing `dynamic` step. This was causing an issue
for the nodes where constraints applied, in other words constraint nodes
are moving. This commit fixes this issue by adding synchronization step
which is done by the `synchronizeBC` introduced in `base_system.py`.
This function contains the list called `_featuresBC` which has the list
of the constraints. `synchronizeBC` function is called every time after
`kinematic` and `dynamic` step are done.