bug fix: np.random number generator usage
This commit fixes a but in `test_intreaction.py` test case. For some
variables, such as normal force we were using `np.random.random_sample`
to generate random number. This function gives you a random number in
array. Thus when you call this function in array, you change your
dimensions of your array and this was causing an issue. With this commit
we are getting the scalar value generated by this random number and set
one element of array with this random number.