bugfix and documentation:`apply_normal_force` func
This commit adds more comments and documentation for
`apply_normal_force` function. As well as we changed some variable names
to prevent confusion. Finally, future possible bug that might occur due
to  missing `()` is fixed. Bug was in where we compute the distance and
surface difference and compare with a tolerance to check if there is
contact.