bug fix:sort problem in constraints and forcing
This commit fixes a bug in constraints and forcing wrappers.
`_ext_forces_torques` and `_constraints` are list of tuples. First index
of tuple is the rod id and second id is type of the constraint or force
applied as string. Previously if you apply two forces or constraints on
the same rod, sort was trying to solve the list with respect to second
id, which is the type of the constraint or force (these are string).
Thus sort could not sort because types were strings. Now we are sorting
with respect to the id of the rod and do not care the type of the
forcing or constraint.