bug fix:PEFRL function fix
This commit fixes a missing and required function call in PEFRL. Before
PEFRL was not calling the `update_internal_forces_and_torques` function
and it was causing test cases to crash with this fix tests will pass.
Also for the analytical solutions we compare in test cases, we need
`update_internal_forces_and_torques` as dummy function.