bug fix:test cases for friction function
This commit fixes a bug for friction test cases. Internal forces and
torques are now variables and we dont use function calls in interaction
module. Thus for the Rod class defined for the test case requires
`internal_forces` and `internal_torques` variables to be defined. In
this commit we are fixing this issue.