bug fix:tangent plane cross product
This commit fixes a bug in `interaction.py`. For the rolling friction,
we require the tangent plane direction and it was computed with cross
product of normal direction with axial direction. However, it should be
axial direction with normal direction. Order was wrong and it is fixed
with this commit.