bug fix:initializing sigma and kappa at beginning
This commit is for a bugfix reported. At the beginning of the
simulation, `sigma` and `kappa` were not initialized. Thus new force
model which requires sigma and kappa at the beginning of the simulation
was not working. We initialize the sigma and kappa arrays at the
beginning of the simulation, by just calling
`_compute_shear_stretch_strains` and `_compute_bending_twist_strains`
functions in `CosseratRod` class constructor.