bug fix: axial direction vector for friction.
This commit fixes an issue for computing axial friction vector
direction. Previously we were setting axial direction and rod element
tangent direction to be the same. However, this is not correct for every
cases. If rod tangent has an out of friction plane component, then it
wont be logical to take tangent as axial friction direction. With this
commit we are projecting tangent vector in to the plane and setting
projected vector as axial direction. Thus friction forces will only
create an external force on rod in plane direction.