bug fix:muscle torque calculation
This commit fixes a bug in `MuscleTorques` class. In elastica cpp
optimized torque profile is set such that head of the snake is first
node and tail of snake is last node. However, in this code we do in
opposite way, thus computed torque profile has to be reordered in order
to obtain correct gaits.