renaming, comments added: _spline.py
This commit adds and changes some of the comments for better code
readability. Also, it updates  and sets coefficients at the head and
tail to be zero, which is the correct case. However, it was not have any
affect on spline, but for preventing misdirecting users, we made this
change.