bug fix: `compute_velocity_center_of_mass` func
This commit fixes a possible future bug. Before we were doing simple
averaging over nodes to compute average velocity of the rod, assuming
that mass of all elements are same. However, this is not the case
always and we can have elements which has different masses. Thus now we
are using mass averaged velocity for calculations.