fix for tests: interaction.py
This commit fixes failed tests for interaction. In
`AnistropicFrictionalPlane` class for computing `axial_direction` in
order to prevent division by 0 we have added a small number  before
(1e-10). However, this was causing tests to be failed. Thus, with this
commit we are decreasing this number to 1e-14. So that assert checks in
tests wont fail.