unused if statement removed in _spline.py
This commit removes unused if statement in`_bspline`. Previously there
was an option which user can give to get back the spline control points.
However, we do not require these points for simulation and they only
might be useful for plotting spline function. Thus with this commit we
remove this option and we plan to increase cod coverage as well.