possible bug fix: interaction.py friction forces
This commit fixes possible future bug in intreaction.py. Friction forces
has to be computed for elements that are in contact with the surface,
for those elements that are not in contact, friction forces has to be
zero. One possible way is to use an if statement and check the
difference between element and surface. However, in python these kind of
if statements will slow the code. Thus, we propose to use `np.where` to
find and store the element ids that are not in contact with the surface
and after computing kinetic or static friction forces for all elements,
we set the friction forces to zero for those elements that are not in
contact, using the element ids. By this way in future, if some elements
are not touching to the surface, we will make sure we are not computing
friction forces for them.