bug fix:`EndpointForcesSinusoidal` force vector
This commit fixes the bug in `EndpointForcesSinusoidal` class force
function. Computed forces are now have same magnitude and sign with the
Elastica.cpp. Only magnitude of the force before ramp up time and sign
of the force components after ramp up time is changed, in order to be
consistent with `Elastica.cpp`.