test connection: fixed
This commit fixes two test cases `test_connection.py` after changing the
connection wrapper. Previously, we were using len to get number of
elements in the rod object in connection wrapper. But since len did not
work. Thus we changed it to copy number of elements from `n_elems`
variable. Because of these changes, we have to fix the `tests` as well.