enhencement:muscle basis functions for octopus
This commit adds, client file for a octopus with two arm. We model the
octopus as a single rod and modify radius of head and arms so that it
mimicks the two arm octopus. We added muscles as basis functions and we
have muscles in each arm replicating 2 bending modes and one twist. We
also changed the post processing files.