enhencement:Rigid body of cylinder is added.
This commit adds rigid cylinder implementation. We also changed and
updated the interaction routines for this cylinder for a specific case,
when it is sitting on top of a plane when its base is facing the plane.
We validated our friction calculations with the analytical solution,
which is in friction validation script.