bugfix:muscle torque function input arguments
This commit fixes an issue in `continuum_snake.py` and
`continuum_flagella.py` cases. Muscle torque function is taking
`rest_lengths` as an input but we were not supplying for the those two
cases. So with this commit we are adding `rest_lengths` as an input
argument for `MuscleTorques` class.