Bugfix: numba jitclass import
This commit fixes a bug which appeared due to the new Numba release. Numba moved jitclass import under the `numba.experimental` from `numba`. As a result it was causing issues during the CI builds and importing elastica modules. In order to solve this we removed the `jitclass` import since it is not used.