bugfix: kinetic friction module
This commit fixes the bug in kinetic friction module for both Numba and Numpy branch. Rolling (transverse) and axial kinetic friction forces uses the weight of rod elements. Previously weight of the element was not projected correctly in rolling and axial directions. This bug could only been seen in snake simulation, because snake is the only case where both friction directions appear for the same problem.