bugfix: sort friction plane end of the forcing list
This commit adds an improvement for force calculations. Friction plane should be the last external forcing class that has to be evaluated. Previously user has to be carefully add friction plane class as a last external forcing class. Now in backhand Elastica sorts forcing classes and pushes the friction class at the end of forcing list.