enhancement: change b-spline implementation
This commit changes the b-spline implementation in code base. B-spline is used to generate muscle torques on continuum snake and flagella. Previously b-spline was forcing end control points to be zero, but with this commit we remove this constrain. Thus, for flagella we added zero control points to the ends to make sure it gives consistent results.