update: remove reflection in finalize.
The reflection written in fianlize is over-engineered and error-prone.
Because of using string-match to find the functionals, the error is
impossible to be backtraced using the debugger. It is conflict of using
mixin, composition and inheritance together: components (forcing, connection,
etc.) should not have a shared functionals, like inheritance, because
mixin class will not have access to the overrided functions.