Fix bug in checking of path for ExportCallback
If any path not equal to a directory is specified, the current implementation throws an error that the export path does not exist. However, it should also be possible to use filenames (instead of just dirnames) such as:
```python
path="dirname/filename"
```
and then the first batch of data should be for example save as `dirname/filename_0.dat.npz`.